Fix warnings

+++ src/splitString.c
@@ -35,7 +35,7 @@ char *toUpper(const char *input)
 {
     char *output = malloc(strlen(input) + 1);
     strcpy(output, input);
-    for (int i = 0; i < strlen(output); i++)
+    for (size_t i = 0; i < strlen(output); i++)
     {
         output[i] = toupper(output[i]);
     }
